change: make tests.common.testcase.Base not inherit from unittest.TestCase
Make tests.common.testcase.Base class does not inherit from
unittest.TestCase to de-couple them and simplify it. And update
*TestCase in the samme module to utilize instance of that Base class.

Because this change brings some major changes of the inheritance
hierarchy and the beharior of TestCase classes. The all of the test
cases for each rules also needs updates to reflect the changes.